+++ automates/text_reading/build.sbt
@@ -5,14 +5,12 @@ organization := "org.clulab"
 scalaVersion := "2.12.4"
 val json4sVersion = "3.5.2"
 
-resolvers += "Artifactory" at "http://artifactory.cs.arizona.edu:8081/artifactory/sbt-release"
 //EclipseKeys.withSource := true
 
 libraryDependencies ++= {
   val procVer = "7.5.1"
 
   Seq(
-//    "org.clulab"    % "glove-840b-300d"           % "0.1.0",
     "org.clulab"    %% "processors-main"          % procVer,
     "org.clulab"    %% "processors-corenlp"       % procVer,
     "org.clulab"    %% "processors-odin"          % procVer,
@@ -32,8 +30,7 @@ libraryDependencies ++= {
     "com.typesafe.play" %% "play-json" % "2.7.0",
     "org.json4s" %% "json4s-jackson" % "0.1.0",
     "org.scala-lang.modules" %% "scala-xml" % "1.0.2",
-    "org.apache.commons" % "commons-text" % "1.4"//,
-//    "org.apache.spark" %% "spark-core" % "3.1.2"
+    "org.apache.commons" % "commons-text" % "1.4"
   )
 }
