+++ automates/text_reading/src/main/resources/application.conf
@@ -80,7 +80,7 @@ apps {
 
   outputDirectory = "./"
 
-  exportAs = ["tsv"]//["serialized", "json"] // no other formats yet specified, but can be added!
+  exportAs = ["tsv", "serialized", "json"] // no other formats yet specified, but can be added!
   loadMentions = true
   mentionsFile = "/local/path/to/PT-mentions.json"
   appendToGrFN = true
@@ -99,15 +99,15 @@ apps {
 
 alignment {
   alignerType = "pairwisew2v"
-  w2vPath = "vectors.txt"   //"org/clulab/glove/glove.840B.300d.txt" // a word embeddings file (e.g., GloVe) with this header: <len_of_vocab><space><embedding_size>, e.g., "6B 50" (no quotes); use of different sets of embeddings will require adjusting alignment similarity thresholds, e.g., commentTextAlignmentScoreThreshold
-  relevantArgs = ["description"]
+  w2vPath = "vectors.txt" // a word embeddings file (e.g., GloVe) with this header: <len_of_vocab><space><embedding_size>, e.g., "6B 50" (no quotes); use of different sets of embeddings will require adjusting alignment similarity thresholds, e.g., commentTextAlignmentScoreThreshold
+  relevantArgs = ["description"] // also possible ["variable"]
 }
 
 // if no other changes come to the model comparison alignment, can remove - tbd
 modelComparisonAlignment {
   alignerType = "pairwisew2v"
-  w2vPath = "vectors.txt" //"/org/clulab/glove/glove.840B.300d.txt"
-  relevantArgs = ["description"]
+  w2vPath = "vectors.txt"
+  relevantArgs = ["description"] // also possible ["variable"]
 }
 
 grounding {
