+++ automates/text_reading/src/main/scala/org/clulab/aske/automates/OdinEngine.scala
@@ -81,7 +81,6 @@ class OdinEngine(
 
     // Run the main extraction engine, pre-populated with the initial state
     val events =  engine.extractFrom(doc, initialState).toVector
-
     //println(s"In extractFrom() -- res : ${res.map(m => m.text).mkString(",\t")}")
     val (descriptionMentions, other) = events.partition(_.label.contains("Description"))
