+++ automates/text_reading/src/main/scala/org/clulab/aske/automates/alignment/Aligner.scala
@@ -17,7 +17,7 @@ import scala.collection.mutable.ArrayBuffer
 case class AlignmentHandler(editDistance: VariableEditDistanceAligner, w2v: PairwiseW2VAligner) {
 
   def this(w2vPath: String, relevantArgs: Set[String]) =
-    this(new VariableEditDistanceAligner(), new PairwiseW2VAligner(new Word2Vec(Sourcer.sourceFromResource(w2vPath), None), relevantArgs))
+    this(new VariableEditDistanceAligner(), new PairwiseW2VAligner(new Word2Vec(w2vPath), relevantArgs))
   def this(config: Config) = this(config[String]("w2vPath"), config[List[String]]("relevantArgs").toSet)
 }
 
@@ -29,13 +29,10 @@ trait Aligner {
   def alignTexts(srcTexts: Seq[String], dstTexts: Seq[String], useBigrams: Boolean): Seq[Alignment]
 }
 
-
 class VariableEditDistanceAligner(relevantArgs: Set[String] = Set("variable"))  {
   def alignMentions(srcMentions: Seq[Mention], dstMentions: Seq[Mention]): Seq[Alignment] = {
     alignTexts(srcMentions.map(Aligner.getRelevantText(_, relevantArgs)), dstMentions.map(Aligner.getRelevantText(_, relevantArgs)))
   }
-
-
 
   def alignTexts(srcTexts: Seq[String], dstTexts: Seq[String]): Seq[Alignment] = {
 
@@ -61,9 +58,7 @@ class VariableEditDistanceAligner(relevantArgs: Set[String] = Set("variable"))  
     if (intersect.length <= 1) return 0
 
     // if intersect contains capital, return len intersect (even with intersect == 1, capital intersect should be informative
-    if (intersect.exists(_.isUpper)) return intersect.length// println("true 75") else println("false 75") //return intersect.length// this is inadequate---has to normalized
-    //println("passed 1")
-
+    if (intersect.exists(_.isUpper)) return intersect.length
 
     // shelving these for now till see more examples; fine for a first pass
     // if intersect == 1, return 1 -> not enough info to make us more confident of the edit distance score
@@ -85,14 +80,13 @@ class VariableEditDistanceAligner(relevantArgs: Set[String] = Set("variable"))  
   }
 
   // todo: maybe for equations it should not be judged on distance but just on overlap in some way? like b and E are in no way similar...
-  // and intersect could be normalized in some way? like by length or something?..
+  // and intersect could be normalized in some way? e.g., by length?
   def alignEqAndTexts(srcTexts: Seq[String], dstTexts: Seq[String]): Seq[Alignment] = {
-    println(srcTexts.mkString("<<++"))
     val exhaustiveScores = for {
       (src, i) <- srcTexts.zipWithIndex
       rendered = AlignmentBaseline.replaceWordWithGreek(AlignmentBaseline.renderForAlign(src), AlignmentBaseline.word2greekDict.toMap)
       (dst, j) <- dstTexts.zipWithIndex
-      multiplier = intersectMultipler(rendered, dst) // how did I make so many eq to gvar elements to go away? is there a threshold?
+      multiplier = intersectMultipler(rendered, dst)
       score = 1.0 * multiplier / (editDistance(rendered, dst) + 1.0) // todo: is this good for long-term? next thing to try: only align if rendered starts with the same letter as actual---might need to make this output an option in case if there are no alignments; todo: try to find a way to use intersect multiplier
     } yield Alignment(i, j, score)
 
@@ -136,7 +130,6 @@ class PairwiseW2VAligner(val w2v: Word2Vec, val relevantArgs: Set[String]) exten
       case x if x.contains("description") => glv.textFromAllDescrs.mkString(" ")
       case _ => ???
     }
-    println("relevant text: " + relText)
     relText
   }
 
@@ -147,30 +140,26 @@ class PairwiseW2VAligner(val w2v: Word2Vec, val relevantArgs: Set[String]) exten
   def alignGlobalCommentVarAndGlobalVars(commentGlVar: Seq[GlobalVariable], glVars: Seq[GlobalVariable]): Seq[Alignment] = {
     // todo: since we base this on text of multiple descriptions, we can try to add some sort of weight for words that occur multiple times
     // todo: text of conj descrs should be returned based on char offset attachement
-    // fixme: for now, edit distance is
-//    for (g <- commentGlVar) println("_> " + g)
     alignTexts(commentGlVar.map(getRelevantTextFromGlobalVar(_).toLowerCase()), glVars.map(getRelevantTextFromGlobalVar(_).toLowerCase()), useBigrams = true)
   }
 
   def alignTexts(srcTexts: Seq[String], dstTexts: Seq[String], useBigrams: Boolean): Seq[Alignment] = {
 
-    for ((src, i) <- srcTexts.zipWithIndex) {
-      for ((dst, j) <- dstTexts.zipWithIndex) {
-        println("++++")
-        println(src.mkString(""))
-        println(dst.mkString(""))
-        println("-> score: " + 1 * compare(src, dst, useBigrams))// + (1.0 / (editDistance(src, dst) + 1.0)))
-      }
-    }
+    // keep for debugging align scores
+//    for ((src, i) <- srcTexts.zipWithIndex) {
+//      for ((dst, j) <- dstTexts.zipWithIndex) {
+//        println("++++")
+//        println(src.mkString(""))
+//        println(dst.mkString(""))
+//        println("-> score: " + 1 * compare(src, dst, useBigrams))// + (1.0 / (editDistance(src, dst) + 1.0)))
+//      }
+//    }
 
-
-
-
+    // fixme: for now, edit distance is disabled---it was dragging the score down because string compared (multiple descriptions from texts and one comment description) ended up being of very different lengths for some global vars
     val exhaustiveScores = for {
       (src, i) <- srcTexts.zipWithIndex
       (dst, j) <- dstTexts.zipWithIndex
       score = compare(src, dst, useBigrams)// + (1.0 / (editDistance(src, dst) + 1.0))
-//      score =  (1.0 / (editDistance(src, dst) + 1.0)) // score without using word embeddings
     } yield Alignment(i, j, score)
     // redundant but good for debugging
     exhaustiveScores
@@ -212,9 +201,7 @@ object PairwiseW2VAligner {
 
     new PairwiseW2VAligner(w2v, relevantArgs.toSet)
   }
-
 }
-
 
 object Aligner {
 
@@ -231,15 +218,13 @@ object Aligner {
       if (debug) println(s"srcIdx: ${srcIdx}, alignments: ${alignments}")
     }
 
-//    val overallSum = alignments.map(_.score).sum
     val grouped = alignments.groupBy(_.dst).toSeq
 
     for {
       (srcIdx, aa) <- grouped
       _ = debugPrint(debug, srcIdx, alignments)
-      // since we added a bunch of multipliers, it might be good to normalize the scores, so we divide them by the sum of all the scores - doesn't work + this is normalization for all the scores to sum up to 1 - not useful
-      // bc scale ends up being different when we choose highest score for dst var
-      scoreMax = aa.map(_.score).max
+      // todo: experiment with score max - divide by max score to account for using weights in alignment
+//      scoreMax = aa.map(_.score).max
       topK = aa.sortBy(-_.score).slice(0,k).filter(_.score > scoreThreshold)
 //      topK = aa.sortBy(-_.score/scoreMax).slice(0,k).filter(_.score > scoreThreshold).map(a => Alignment(a.src, a.dst, a.score/scoreMax))//filter out those with the score below the threshold; threshold is 0 by default
       if topK.nonEmpty
@@ -250,21 +235,12 @@ object Aligner {
     def debugPrint(debug: Boolean, srcIdx: Int, alignments: Seq[Alignment]) {
       if (debug) println(s"srcIdx: ${srcIdx}, alignments: ${alignments}")
     }
-
-//    val overallSum = alignments.map(_.score).sum
     val grouped = alignments.groupBy(_.src).toSeq
 
-    // score sum doesn't work bc alignment for eq to text bc here we group by source and src is eq and then when we choose top n for dst,
-    // if we normalize by sum, we end up on a different scale
-//    for ((srcIdx, aa) <- grouped) {
-//      val scoreSum = aa.map(_.score).sum
-//      println("score sum: " + scoreSum)
-//    }
+    // todo: experiment with score max - divide by max score to account for using weights in alignment; can't do that by sum because that is more like argmax
     for {
       (srcIdx, aa) <- grouped
       _ = debugPrint(debug, srcIdx, alignments)
-      // since we added a bunch of multipliers, it might be good to normalize the scores, so we divide them by the sum of all the scores - doesn't work
-      // bc scale ends up being different when we choose highest score for dst var
       scoreMax = aa.map(_.score).max // if i do weights for some of the links, then when getting top k, divide sort by score by score max and remap alignments to updated scores
       topK = aa.sortBy(-_.score).slice(0,k).filter(_.score > scoreThreshold)//.map(a => Alignment(a.src, a.dst, a.score/scoreMax)) //filter out those with the score below the threshold; threshold is 0 by default
       if topK.nonEmpty
