+++ automates/text_reading/src/test/resources/test.conf
@@ -54,7 +54,7 @@ apps {
 
 alignment {
   debug = true // for align test, always has to be true
-  w2vPath = "/vectors.txt" ///org/clulab/glove/glove.840B.300d.txt"
+  w2vPath = "vectors.txt" ///org/clulab/glove/glove.840B.300d.txt"
   unitTestPayload = "align_payload-for-testing.json"
   fullLinkEvalPayload = "LinkEval/align_payload.json"
 }
